requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
black==26.1.0
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
fastapi==0.128.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
mypy_extensions==1.1.0
openai==2.15.0
packaging==26.0
pathspec==1.0.3
pillow==12.1.0
platformdirs==4.5.1
pydantic==2.12.5
pydantic_core==2.41.5
pytesseract==0.3.13
python-dotenv==1.2.1
python-multipart==0.0.21
pytokens==0.4.0
ruff==0.14.14
sniffio==1.3.1
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
